tests/wpt/web-platform-tests/offscreen-canvas/fill-and-stroke-styles/2d.strokeStyle.default.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by tools/gentest.py. -->
<title>OffscreenCanvas test: 2d.strokeStyle.default</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/canvas-tests.js"></script>

<h1>2d.strokeStyle.default</h1>
<p class="desc"></p>


<script>
var t = async_test("");
t.step(function() {

var offscreenCanvas = new OffscreenCanvas(100, 50);
var ctx = offscreenCanvas.getContext('2d');

_assertSame(ctx.strokeStyle, '#000000', "ctx.strokeStyle", "'#000000'");

t.done();

});
</script>
